+++ ChangeLog.md
@@ -10,6 +10,7 @@
 
 ### Changed
 - 事件全部使用`类`来描述
+- ConsoleAppSettingInterface::getCmdMapping更名为ConsoleAppSettingInterface::getCmds
 
 ### Removed
 - 去掉了“filter_field”、“index2map”函数，使用PHP自带的“array_column”代替